+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/process/traversal/step/map/AddEdgeStepContract.java
@@ -19,11 +19,9 @@
 package org.apache.tinkerpop.gremlin.process.traversal.step.map;
 
 import java.util.List;
-import java.util.function.Consumer;
 import org.apache.tinkerpop.gremlin.process.traversal.Traversal;
 import org.apache.tinkerpop.gremlin.process.traversal.Traverser;
 import org.apache.tinkerpop.gremlin.process.traversal.lambda.ConstantTraversal;
-import org.apache.tinkerpop.gremlin.process.traversal.lambda.GValueConstantTraversal;
 import org.apache.tinkerpop.gremlin.process.traversal.step.FromToModulating;
 import org.apache.tinkerpop.gremlin.process.traversal.step.GValue;
 import org.apache.tinkerpop.gremlin.process.traversal.step.Scoping;
@@ -36,8 +34,38 @@ import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.apache.tinkerpop.gremlin.structure.util.reference.ReferenceVertex;
 
 public interface AddEdgeStepContract<S> extends TraversalParent, Scoping, FromToModulating, AddElementStepContract<S, Edge>, Writing<Event.EdgeAddedEvent> {
+
+    /**
+     * Gets the "from" vertex for the edge to be added. If the "from" vertex was added as a {@link Vertex}, ID, {@link GValue},
+     * or {@link ConstantTraversal}, it is returned as a {@link ReferenceVertex}. Otherwise, it is returned in {@link Traversal} form.
+     */
     Object getFrom();
+
+    /**
+     * Gets the "to" vertex for the edge to be added. If the "to" vertex was added as a {@link Vertex}, ID, {@link GValue},
+     * or {@link ConstantTraversal}, it is returned as a {@link ReferenceVertex}. Otherwise, it is returned in {@link Traversal} form.
+     */
     Object getTo();
+
+    /**
+     * Gets the "from" vertex for the edge to be added. If the "from" vertex was added as a {@link Vertex}, ID,
+     * or {@link ConstantTraversal}, it is returned as a {@link ReferenceVertex}. If it was added as a {@link GValue}
+     * containing a {@link Vertex} or ID, the {@link GValue} is returned. Otherwise, it is returned in {@link Traversal} form.
+     */
+    default Object getFromWithGValue() {
+        Object from = getFrom();
+        return from instanceof Traversal ? from : GValue.of(from);
+    }
+
+    /**
+     * Gets the "from" vertex for the edge to be added. If the "from" vertex was added as a {@link Vertex}, ID,
+     * or {@link ConstantTraversal}, it is returned as a {@link ReferenceVertex}. If it was added as a {@link GValue}
+     * containing a {@link Vertex} or ID, the {@link GValue} is returned. Otherwise, it is returned in {@link Traversal} form.
+     */
+    default Object getToWithGValue() {
+        Object to = getTo();
+        return to instanceof Traversal ? to : GValue.of(to);
+    }
 
     default Object getAdjacentVertex(Parameters parameters, String key) {
         Object vertex;
@@ -72,21 +100,5 @@ public interface AddEdgeStepContract<S> extends TraversalParent, Scoping, FromTo
         return (Vertex) vertex;
     }
 
-    default Object resolveVertexTraversal(Traversal.Admin<?,?> vertexTraversal) {
-        return resolveVertexTraversal(vertexTraversal, null);
-    }
 
-    default Object resolveVertexTraversal(Traversal.Admin<?,?> vertexTraversal, Consumer<GValue<?>> gValueConsumer) {
-        if (vertexTraversal instanceof GValueConstantTraversal) {
-            if (gValueConsumer != null) {
-                gValueConsumer.accept(((GValueConstantTraversal<?, ?>) vertexTraversal).getGValue());
-            }
-            return new ReferenceVertex(vertexTraversal.next());
-        }
-        if (vertexTraversal instanceof ConstantTraversal) {
-            return new ReferenceVertex(vertexTraversal.next());
-        }
-        // anything other than a ConstantTraversal is returned as-is
-        return vertexTraversal;
-    }
 }
