+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/process/traversal/step/map/AddEdgeStepPlaceholder.java
@@ -18,19 +18,13 @@
  */
 package org.apache.tinkerpop.gremlin.process.traversal.step.map;
 
-import java.util.Collection;
 import org.apache.tinkerpop.gremlin.process.traversal.Traversal;
 import org.apache.tinkerpop.gremlin.process.traversal.lambda.ConstantTraversal;
 import org.apache.tinkerpop.gremlin.process.traversal.lambda.GValueConstantTraversal;
 import org.apache.tinkerpop.gremlin.process.traversal.step.GValue;
-import org.apache.tinkerpop.gremlin.process.traversal.step.util.event.Event;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 
-public class AddEdgeStepPlaceholder<S> extends AbstractAddElementStepPlaceholder<S, Edge, Event.EdgeAddedEvent>
-        implements AddEdgeStepContract<S> {
-
-    private Traversal.Admin<?,?> from;
-    private Traversal.Admin<?,?> to;
+public class AddEdgeStepPlaceholder<S> extends AbstractAddEdgeStepPlaceholder<S> {
 
     public AddEdgeStepPlaceholder(final Traversal.Admin traversal, final String edgeLabel) {
         super(traversal, edgeLabel == null ? Edge.DEFAULT_LABEL : edgeLabel);
@@ -46,36 +40,6 @@ public class AddEdgeStepPlaceholder<S> extends AbstractAddElementStepPlaceholder
     }
 
     @Override
-    public void addTo(final Traversal.Admin<?, ?> toObject) {
-        addTraversal(toObject);
-        if (toObject instanceof GValueConstantTraversal) {
-            traversal.getGValueManager().register(((GValueConstantTraversal<?, ?>) toObject).getGValue());
-        }
-        this.to = toObject;
-    }
-
-    @Override
-    public void addFrom(final Traversal.Admin<?, ?> fromObject) {
-        addTraversal(fromObject);
-        if (fromObject instanceof GValueConstantTraversal) {
-            traversal.getGValueManager().register(((GValueConstantTraversal<?, ?>) fromObject).getGValue());
-        }
-        this.from = fromObject;
-    }
-
-    @Override
-    public int hashCode() {
-        int hash = super.hashCode();
-        if (from != null) {
-            hash ^= from.hashCode();
-        }
-        if (to != null) {
-            hash ^= to.hashCode();
-        }
-        return hash;
-    }
-
-    @Override
     public AddEdgeStep<S> asConcreteStep() {
         AddEdgeStep<S> step = new AddEdgeStep<>(traversal, label instanceof GValueConstantTraversal ? ((GValueConstantTraversal<S, String>) label).getConstantTraversal() : label);
         super.configureConcreteStep(step);
@@ -88,71 +52,4 @@ public class AddEdgeStepPlaceholder<S> extends AbstractAddElementStepPlaceholder
         return step;
     }
 
-    @Override
-    public boolean isParameterized() {
-        if (super.isParameterized() ||
-                (from instanceof GValueConstantTraversal && ((GValueConstantTraversal<S, String>) from).isParameterized()) ||
-                (to instanceof GValueConstantTraversal && ((GValueConstantTraversal<S, String>) to).isParameterized())) {
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    protected boolean supportsMultiProperties() {
-        return false;
-    }
-
-    @Override
-    public Object getFrom() {
-        return resolveVertexTraversal(from, gValue -> traversal.getGValueManager().pinVariable(gValue.getName()));
-    }
-
-    public Object getFromGValueSafe() {
-        return resolveVertexTraversal(from);
-    }
-
-    @Override
-    public Object getTo() {
-        return resolveVertexTraversal(to, gValue -> traversal.getGValueManager().pinVariable(gValue.getName()));
-    }
-
-    public Object getToGValueSafe() {
-        return resolveVertexTraversal(to);
-    }
-
-    @Override
-    public void updateVariable(String name, Object value) {
-        super.updateVariable(name, value);
-        if (from instanceof GValueConstantTraversal) {
-            ((GValueConstantTraversal<S, String>) from).updateVariable(name, value);
-        }
-        if (to instanceof GValueConstantTraversal) {
-            ((GValueConstantTraversal<S, String>) to).updateVariable(name, value);
-        }
-    }
-
-    @Override
-    public Collection<GValue<?>> getGValues() {
-        Collection<GValue<?>> gValues = super.getGValues();
-        if (from instanceof GValueConstantTraversal && ((GValueConstantTraversal<S, String>) from).getGValue().isVariable()) {
-            gValues.add(((GValueConstantTraversal<S, String>) from).getGValue());
-        }
-        if (to instanceof GValueConstantTraversal && ((GValueConstantTraversal<S, String>) to).getGValue().isVariable()) {
-            gValues.add(((GValueConstantTraversal<S, String>) to).getGValue());
-        }
-        return gValues;
-    }
-
-    @Override
-    public AddEdgeStepPlaceholder<S> clone() {
-        final AddEdgeStepPlaceholder<S> clone = (AddEdgeStepPlaceholder) super.clone();
-        if (from != null){
-            clone.from = from.clone();
-        }
-        if (to != null){
-            clone.to = to.clone();
-        }
-        return clone;
-    }
 }
